requirements.txt
streamlit>=1.55
pandas>=2.0
psycopg2-binary>=2.9
anthropic>=0.40
python-dotenv>=1.2
